% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{cov2dist}
\alias{cov2dist}
\title{cov2dist function}
\usage{
cov2dist(XtX)
}
\arguments{
\item{XtX}{Cross-product (\eqn{X'X}) of a matrix \eqn{X} containing \eqn{p} (centered) variables.
If a variance-covariance matrix is provided as \eqn{X'X/(n-1)} then the output is \eqn{D/(n-1)}.}
}
\value{
A squared matrix D containing the squared Euclidean distances
}
\description{
Computes a squared Euclidean distance matrix from a covariance matrix among \eqn{p} variables. The pairwise distance \eqn{d(x,y)} between 
variables \eqn{x} and \eqn{y} is obtained from the cross-product of the vectors \eqn{\textbf{x}'\textbf{y}=\sum{x_iy_i}} as
\deqn{d^2(x,y)=(\textbf{x}-\textbf{y})'(\textbf{x}-\textbf{y})=\textbf{x}'\textbf{x} + \textbf{y}'\textbf{y} - 2\textbf{x}'\textbf{y}}
If the variables are centered then \eqn{\textbf{x}'\textbf{y}/(n-1)=cov(x,y)} is the sample covariance. Then
\deqn{d^2(x,y)/(n-1)=var(x) + var(y) - 2cov(x,y)}
}
\examples{
set.seed(1234)
require(SFSI)
# Simulate matrix
n = 100; p=10
X = scale(matrix(rnorm(n*p),ncol=p))

# Distance matrix from a cross-product
COV = crossprod(X)   # Cross-product X'X
cov2dist(COV)  # equal to: as.matrix(dist(t(X)))^2

# Distance matrix from a variance-covariance matrix
COV = var(X)   # Variance matrix of X
(n-1)*cov2dist(COV)  # equal to: as.matrix(dist(t(X)))^2
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
