% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{getDistantSet}
\alias{getDistantSet}
\title{getDistantSet function}
\usage{
getDistantSet(G = NULL, PC = NULL, n, kernel = NULL)
}
\arguments{
\item{G}{Relationship matrix among individuals}

\item{PC}{Principal components from the relationship matrix 'G'. When 'PC' is provided the 'G' is not needed}

\item{n}{Number of individuals to select}

\item{kernel}{Kernel transformation applied to 'G'. List consisting on one of:
\itemize{
  \item list(kernel='GAUSSIAN',h). If \eqn{h} is not provided the value of \eqn{h=-2*log(0.5)} is used.
  \item list(kernel='LAPLACIAN',h). If \eqn{h} is not provided the value of \eqn{h=-2*log(0.5)} is used.
  \item list(kernel='POLYNOMIAL',a,b). The values of \eqn{a=1} and \eqn{b=2} are used when they are not provided.
}
Default kernel=NULL (no kernel)}
}
\value{
A vector with the index of the maximally distant individuals
}
\description{
Selects a set of \eqn{n} individuals that are maximally distant. The individuals are selected as the closest ones
to the centroid of each cluster that are generated using k-means clustering using the principal components of a relationship matrix.
}
