% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{Methods_SFI_CV}
\alias{summary.SFI_CV}
\alias{plot.SFI_CV}
\title{SFI_CV methods}
\usage{
\method{summary}{SFI_CV}(object, ...)

\method{plot}{SFI_CV}(..., py=c("accuracy","MSE"), title=NULL, showFolds=FALSE)
}
\arguments{
\item{object}{An object of the class 'SFI_CV'}

\item{...}{Arguments to be passed: 
\itemize{
 \item Not needed for \code{summary} function
 \item \code{object}: One or more objects of the class 'SFI_CV' (for function \code{plot})
}
}

\item{py}{Indicates whether to plot correlation (between observed and predicted) or MSE in y-axis}

\item{title}{Title of the plot}

\item{showFolds}{\code{TRUE} or \code{FALSE} to whether add results for individuals folds}
}
\description{
\itemize{
 \item \code{'summary'}: Returns accuracy, MSE, df, achieved by each value of the vector lambda. These values are averages across folds and partitions. Also returns summary for the SFI with maximum accuracy and with minimum MSE.
 \item \code{'plot'}: Creates a plot of either accuracy or MSE versus the (average) number of predictors and versus lambda (in negative logarithm) averaged across folds and partitions.
}
}
\examples{
  require(SFSI)
  data(wheat.E1) 
  data(wheat.G) # Genomic relationship matrix
  y = tapply(Y[,"YLD"],Y[,"gid"],mean)  # Response variable
  index <- intersect(rownames(G),names(y))
  G = G[index,index]
  y = as.vector(y[index])
  
  fm1 = SFI_CV(y,K=G,trn.CV=1:500,nFolds=5,name="CV5")
  fm2 = SFI_CV(y,K=G,trn.CV=1:500,nFolds=10,name="CV10")
  
  yHat = fitted(fm1)        # Predicted values for each SFI
  corTST = summary(fm1)$accuracy  # Testing set accuracy (correlation cor(y,yHat))
  summary(fm1)$optCOR       # SFI with maximum accuracy
  summary(fm1)$optMSE       # SFI with minimum MSE
  plot(fm1,title=expression('corr('*y[obs]*','*y[pred]*') vs sparsity'))   
  plot(fm1,py="MSE",title='Mean Square Error vs sparsity')     
  plot(fm2,showFolds=TRUE)     
  plot(fm1,fm2)   
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
