% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{scale_crossprod}
\alias{scale_crossprod}
\title{scale_crossprod function}
\usage{
scale_crossprod(XtX)
}
\arguments{
\item{XtX}{Variance-covariance matrix among \eqn{p} variables.}
}
\value{
A list object containing the elements:
\itemize{
  \item \code{XtX}: squared matrix with the recalculated variances and covariances
  \item \code{sdX}: vector containing the standard deviations of each variable
}
}
\description{
Recalculate the variance-covariance matrix among \eqn{p} variables with unit variance from a variance-covariance matrix where the 
variables have are no neccesarily variance equal to one. The new matrix will contain as off-diagonal entries the covariance between 
the scaled variables \eqn{x^\star=x/\sigma_x} and \eqn{y^\star=y/\sigma_y}, formed by dividing original variables \eqn{x} and \eqn{y} 
by their standard deviation \eqn{\sigma_x} and \eqn{\sigma_y}, given by
\deqn{cov(x^\star,y^\star)=cov(x,y)/(\sigma_x\sigma_y)}
while in the diagonal the variance will be \eqn{var(x^\star)=var(x)/\sigma^2_x=1}.
}
\examples{
set.seed(1234)
require(SFSI)
# Simulate matrix
n = 100; p=10
X = matrix(rnorm(n*p),ncol=p)
COV = var(X)   # Cross-product X'X

COV2 = scale_crossprod(COV)
COV2$XtX   # equal to: var(scale(X))
COV2$sdX   # equal to: apply(X,2,sd)
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
