% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBLUP.R
\name{solveMixed}
\alias{solveMixed}
\title{solveMixed function}
\usage{
solveMixed(y, X = NULL, Z = NULL, K, indexK = NULL, h2 = NULL,
     BLUE = TRUE, BLUP = TRUE, method = "ML", return.Hinv = FALSE,
     tol = 1E-5, maxIter = 1000, interval = c(1E-10,1E10))
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix for the fixed effects. When \code{X=NULL} a vector of ones is constructed only for the intercept (default)}

\item{Z}{Design matrix for the random effects. When \code{Z=NULL} an identity matrix is considered (default) thus \code{G = Z}; otherwise \code{G = Z\%*\%K\%*\%t(Z)} is used}

\item{K}{Kinship relationships matrix. This can be a name of a binary file where the matrix is storaged}

\item{indexK}{Vector of integers indicating which columns and rows will be read when \code{K} is the name of a binary file. Default \code{indexK=NULL} will read the whole matrix}

\item{h2}{Heritability of the response variable. When \code{h2=NULL}, the heritability is calculated from variance components estimated using Maximum Likelihood (ML)}

\item{BLUE}{\code{TRUE} or \code{FALSE} to whether return the fixed effects estimates}

\item{BLUP}{\code{TRUE} or \code{FALSE} to whether return the random effects estimates}

\item{method}{Either 'ML' (Maximum Likelihood) or 'REML' (Restricted Maximum Likelihood). Only 'ML' method is implemented in this version}

\item{return.Hinv}{\code{TRUE} or \code{FALSE} to whether return the inverse of the matrix H}

\item{tol}{Maximum error between two consecutive solutions when solving the root}

\item{maxIter}{Maximum number of iterations to run before convergence is reached}

\item{interval}{Range of values in which the root is searched}
}
\description{
Solves the Linear Mixed Model and calculates Best Linear Unbiased Predictor (BLUP)
}
\details{
The basic linear mixed model that relates phenotypes with genetic values is of the form

\ifelse{html}{\out{<center><b>y</b> = <b>X b</b> + <b>Z g</b> + <b>e</b></center>}}{\deqn{\textbf{y}=\textbf{X}\textbf{b}+\textbf{Z}\textbf{g}+\textbf{e}}{y = X b + Z g + e}}

where
\ifelse{html}{\out{<b>y</b>}}{\eqn{\textbf{y}}{y}} is a vector with the response,
\ifelse{html}{\out{<b>b</b>}}{\eqn{\textbf{b}}{b}} is the vector of fixed effects,
\ifelse{html}{\out{<b>g</b>}}{\eqn{\textbf{u}}{u}} is the vector of the genetic values of the genotypes,
\ifelse{html}{\out{<b>e</b>}}{\eqn{\textbf{e}}{e}} is the vector of environmental residuals, and
\ifelse{html}{\out{<b>X</b>}}{\eqn{\textbf{X}}{X}} and \ifelse{html}{\out{<b>Z</b>}}{\eqn{\textbf{Z}}{Z}} are design matrices conecting the fixed and genetic effects with replicates. Genetic values are assumed to follow a Normal distribution as
\ifelse{html}{\out{<b>g</b> ~ N(<b>0</b>,&sigma;<sup>2</sup><sub>u</sub><b>K</b>)}}{\eqn{\textbf{u}\sim N(\textbf{0},\sigma^2_u\textbf{K})}{u ~ N(0,sigma^2_u K)}}, and environmental terms are assumed
\ifelse{html}{\out{<b>e</b> ~ N(<b>0</b>,&sigma;<sup>2</sup><sub>e</sub><b>I</b>)}}{\eqn{\textbf{e}\sim N(\textbf{0},\sigma^2_e\textbf{I})}{u ~ N(0,sigma^2_e I)}}.

The resulting vector of genetic values
\ifelse{html}{\out{<b>u</b> = <b>Z g</b>}}{\eqn{\textbf{u}=\textbf{Z}\textbf{g}}{u = Z g}} will therefore follow 
\ifelse{html}{\out{<b>u</b> ~ N(<b>0</b>,&sigma;<sup>2</sup><sub>u</sub><b>G</b>)}}{\eqn{\textbf{u}\sim N(\textbf{0},\sigma^2_u\textbf{G})}{u ~ N(0,sigma^2_u G)}} where
\ifelse{html}{\out{<b>G</b> = <b>Z K Z'</b>}}{\eqn{\textbf{G}=\textbf{Z}\textbf{K}\textbf{Z}'}{G = Z K Z'}}.
In the un-replicated case, \ifelse{html}{\out{<b>Z</b> = <b>I</b>}}{\eqn{\textbf{Z}=\textbf{I}}{Z = I}} is an identity matrix, and hence 
\ifelse{html}{\out{<b>u</b> = <b>g</b>}}{\eqn{\textbf{u}=\textbf{g}}{u = g}} and
\ifelse{html}{\out{<b>G</b> = <b>K</b>}}{\eqn{\textbf{G}=\textbf{K}}{G = K}}.

The values \ifelse{html}{\out{<b>u</b><sub>tst</sub> = (u<sub>i</sub>)}}{\eqn{\textbf{u}_{tst}=(u_i)}{u_tst = (u_i)}},
\ifelse{html}{\out{i = 1,2,...,n<sub>tst</sub>}}{\eqn{i=1,2,...,n_{tst}}{i = 1,2,...,n_tst}}, for a testing set are estimated using (as predictors) all available observations in a training set as

\ifelse{html}{\out{<center><b>u</b><sub>tst</sub> = <b>H</b> (<b>y</b><sub>trn</sub> - <b>X</b><sub>trn</sub><b>b</b>)</center>}}{\deqn{\textbf{u}_{tst}=\textbf{H}(\textbf{y}_{trn}-\textbf{X}_{trn}\textbf{b})}{u_tst = H (y_trn - X_trn*b)}}

where \ifelse{html}{\out{<b>H</b>}}{\eqn{\textbf{H}}{H}}
is a matrix of weights given by
\ifelse{html}{\out{<center><b>H</b> = <b>G</b><sub>tst,trn</sub> (<b>G</b><sub>trn,trn</sub> + &lambda;<sub>0</sub><b>I</b>)<sup>-1</sup></center>}}{\deqn{\textbf{H}=\textbf{G}_{tst,trn}(\textbf{G}_{trn,trn} + \lambda_0\textbf{I})^{-1}}{H = G[tst,trn](G[trn,trn] + lambda_0 I)^-1}}

where
\ifelse{html}{\out{<b>G</b><sub>tst,trn</sub>}}{\eqn{\textbf{G}_{tst,trn}}{G[tst,trn]}}
is the sub-matrix of \ifelse{html}{\out{<b>G</b>}}{\eqn{\textbf{G}}{G}} whose rows correspond to the testing set and columns to the training set,
\ifelse{html}{\out{<b>G</b><sub>trn,trn</sub>}}{\eqn{\textbf{G}_{trn,trn}}{G[trn,trn]}}
is the sub-matrix corresponding to the training set, and \ifelse{html}{\out{&lambda;<sub>0</sub> = (1 - h<sup>2</sup>)/h<sup>2</sup>}}{\eqn{\lambda_0=(1-h^2)/h^2}{lambda_0 = (1 - h^2)/h^2}} is a shrinkage parameter expressed in terms of the heritability, \ifelse{html}{\out{h<sup>2</sup> = &sigma;<sup>2</sup><sub>u</sub>/(&sigma;<sup>2</sup><sub>u</sub> + &sigma;<sup>2</sup><sub>e</sub>)}}{\eqn{h^2=\sigma^2_u/(\sigma^2_u+\sigma^2_e)}{h^2 = sigma^2_u/(sigma^2_u + sigma^2_e)}}.
}
\examples{
  require(SFSI)
  data(wheat.E1) 
  data(wheat.G) # Genomic relationship matrix
  y = tapply(Y[,"YLD"],Y[,"gid"],mean)  # Response variable
  index <- intersect(rownames(G),names(y))
  G = G[index,index]
  y = as.vector(y[index])

  # Training and testing sets
  tst = sample(seq_along(y),ceiling(0.3*length(y)))
  trn = seq_along(y)[-tst]

  yNA <- y
  yNA[tst] <- NA
  fm = solveMixed(yNA,K=G)
  plot(y[tst],fm$u[tst])      # Predicted vs observed values in testing set
  cor(y[tst],fm$u[tst])       # Prediction accuracy in testing set
  cor(y[trn],fm$u[trn])       # Prediction accuracy in training set
  fm$h2                       # Heritability
}
\references{
\itemize{
\item \insertRef{VanRaden2008}{SFSI}
}
}
\author{
Paulino Perez, Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
\keyword{solveMixed}
