\name{Y}
\alias{Y}
\title{Wheat phenotypic data}
\description{

A matrix containing traits grain yield (YLD), days to heading (DTH), days to maturity (DTM), and plant height (PH).

Measurements of YLD was reported as the total plot yield after maturity for 1092 genotypes, divided into 39 trials, 3 replicates, and 6 sub-blocks (within replicate). Records for YLD are reported as adjusted phenotypes from which trial, replicate and sub-block effects were removed. The resulting phenotype is the sum of a genotype and an environmental effects estimated from linear mixed models.

Measurements for DTH, DTM, and PH were recorded only in the first replicate at each trial and thus no phenotype adjustment was made.

The number of unique genotypes and number of total records in each environment is
\itemize{
\item Environment 1 (Flat-Drought):    1092 genotypes, 3274 records
\item Environment 2 (Bed-2IR):         1090 genotypes, 3267 records
\item Environment 3 (Bed-5IR):         1092 genotypes, 3274 records
\item Environment 4 (Bed-EHeat):       1013 genotypes, 2858 records
}
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\keyword{datasets}