% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{cov2dist}
\alias{cov2dist}
\title{cov2dist function}
\usage{
cov2dist(XtX)
}
\arguments{
\item{XtX}{Cross-product (\eqn{X'X}) of a matrix \eqn{X} containing \eqn{n} observations and \eqn{p} (centered) variables.
If a variance-covariance matrix is provided as \eqn{X'X/(n-1)} then the output is \eqn{D/(n-1)}.}
}
\value{
A squared matrix D containing the squared Euclidean distances
}
\description{
Computes the pairwise squared Euclidean distance \eqn{d(x,y)} from the covariance between variables 'x' and 'y'
given by the cross-product \eqn{x'y}
\deqn{d^2(x,y)=(x-y)'(x-y)=x'x + y'y - 2x'y}
If the vectors 'x' and 'y' are centered then x'y/(n-1)=cov(x,y) is the sample covariance, where \eqn{n} is the length of the vectors. Then
\deqn{d^2(x,y)/(n-1)=var(x) + var(y) - 2cov(x,y)}
}
\examples{
set.seed(1234)
require(SFSI)
# Simulate matrix
n = 100; p=10
X = scale(matrix(rnorm(n*p),ncol=p))

# Distance matrix from a cross-product
COV = crossprod(X)   # Cross-product X'X
cov2dist(COV)  # equal to: as.matrix(dist(t(X)))^2

# Distance matrix from a variance-covariance matrix
COV = var(X)   # Variance matrix of X
(n-1)*cov2dist(COV)  # equal to: as.matrix(dist(t(X)))^2
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
