% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBLUP.R
\name{GBLUP}
\alias{GBLUP}
\title{GBLUP function}
\usage{
GBLUP(G, y, h2 = 0.5, trn = 1:length(y), tst = 1:length(y),
  kernel = NULL)
}
\arguments{
\item{G}{Genetic relatedness matrix}

\item{y}{Response variable}

\item{h2}{Heritability of the response variable. Default is \code{h2=0.5}}

\item{trn}{Vector of integers indicating which individuals are in training set. Default is \code{trn=1:length(y)} will consider all individuals as training}

\item{tst}{Vector of integers indicating which individuals are in testing set. Default is \code{tst=1:length(y)} will consider all individuals as testing}

\item{kernel}{Kernel transformation to be applied to \code{G[trn,trn]}. See \code{help(kernel2)} for details. Default \code{kernel=NULL} (no kernel)}
}
\description{
Fits the kinship-based Best Linear Unbiased Prediction (BLUP) model
}
\details{
Finds solutions for the genetic values of individuals in a linear model

\ifelse{html}{\out{<center>y<sub>i</sub> = u<sub>i</sub> + e<sub>i</sub></center>}}{\deqn{y_i=u_i+e_i}{y_i = u_i + e_i}}

where
\ifelse{html}{\out{y<sub>i</sub>}}{\eqn{y_i}{y_i}} is the response for the \ifelse{html}{\out{i<sup>th</sup>}}{\eqn{i^{th}}{i^th}} observation,
\ifelse{html}{\out{u<sub>i</sub>}}{\eqn{u_i}{u_i}}
is the genetic value of the individual, and
\ifelse{html}{\out{e<sub>i</sub>}}{\eqn{e_i}{e_i}}
is an environmental residual.

The values \ifelse{html}{\out{u<sub>i</sub>}}{\eqn{u_i}{u_i}} for
\ifelse{html}{\out{i = 1,2,...,n<sub>tst</sub>}}{\eqn{i=1,2,...,n_{tst}}{i = 1,2,...,n_tst}} individuals in a testing set are estimated
using (as predictors) all available observations in a training set as

\ifelse{html}{\out{<center>u<sub>i</sub> = <b>y</b>'<sub>trn</sub> <b>&beta;</b><sub>i</sub></center>}}{\deqn{u_i=\textbf{y}_{trn}'\boldsymbol{\beta}_i}{u_i = y'_trn beta_i}}

where \ifelse{html}{\out{<b>&beta;</b><sub>i</sub>}}{\eqn{\boldsymbol{\beta}_i}{beta_i}}
is a vector of regression coefficients which are estimated as function of the genetic covariance between predictors
and that individual (\ifelse{html}{\out{<b>G</b><sub>trn,tst(i)</sub>}}{\eqn{\textbf{G}_{trn,tst(i)}}{G[trn,tst(i)]}}) and the genetic variance among predictors (\ifelse{html}{\out{<b>G</b><sub>trn,trn</sub>}}{\eqn{\textbf{G}_{trn,trn}}{G[trn,trn]}}).

These coefficients are obtained, separately for each individual in the testing set, by
\ifelse{html}{\out{<center><b>&beta;</b><sub>i</sub> = (<b>G</b><sub>trn,trn</sub> + &lambda;<sub>0</sub><b>I</b>)<sup>-1</sup> <b>G</b><sub>trn,tst(i)</sub></center>}}{\deqn{\boldsymbol{\beta}_i=(\textbf{G}_{trn,trn} + \lambda_0\textbf{I})^{-1}\textbf{G}_{trn,tst(i)}}{beta_i = (G[trn,trn] + lambda_0 I)^-1 G[trn,tst(i)]}}

where \ifelse{html}{\out{&lambda;<sub>0</sub> = (1-h<sup>2</sup>)/h<sup>2</sup>}}{\eqn{\lambda_0=(1-h^2)/h^2}{lambda_0 = (1-h^2)/h^2}} is a parameter expressed in terms of the heritability.
}
\examples{
  \dontrun{
  set.seed(1234)
  require(SFSI)
  # Read data from BGLR package
  data(wheat,package="BGLR")
  X = scale(wheat.X)
  G = tcrossprod(X)/ncol(X)    # Genomic relationship matrix
  y = wheat.Y[,1]              # Response variable

  # Training and testing sets
  set.seed(1234)
  tst = sample(seq_along(y),150)     # 150 individuals to predict
  trn = seq_along(y)[-tst]           # 449 individuals to train model

  # Calculate heritability
  fm = BGLR::BGLR(y,ETA=list(list(K=G,model="RKHS")),nIter=10000,burnIn=3000,verbose=FALSE)
  varU = fm$ETA[[1]]$varU
  varE = fm$varE
  h2 = varU/(varU + varE)

  fm = GBLUP(G,y,h2,trn,tst)
  plot(y[tst],fitted(fm))        # Predicted vs observed values in testing set
  predict(fm)$correlation        # Testing set accuracy
  }
}
\references{
\itemize{
\item \insertRef{Perez2014}{SFSI}
\item \insertRef{VanRaden2008}{SFSI}
}
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
\keyword{GBLUP}
