% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lars.R
\name{lars2}
\alias{lars2}
\title{Least Angle Regression to solve the LASSO-type problem}
\usage{
lars2(P, cov, method = c("LAR", "LAR-LASSO"), maxDF = NULL,
  eps = .Machine$double.eps, scale = TRUE, verbose = FALSE)
}
\arguments{
\item{P}{Variance-covariance matrix among predictors}

\item{cov}{Covariance vector between response variable and predictors}

\item{method}{One of:
\itemize{
 \item \code{'LAR'}: Computes the entire sequence of all coefficients. Values of lambdas are calculated at each step.
 \item \code{'LAR-LASSO'}: Similar to \code{'LAR'} but solutions when a predictor leaves the solution are also returned.
}
Default is \code{method='LAR'}}

\item{maxDF}{Maximum number of predictors in the last lars solution.
Default \code{maxDF=NULL} will calculate solution for all the predictors}

\item{eps}{An effective zero. Default is the machine precision}

\item{scale}{\code{TRUE} or \code{FALSE} to recalculate the matrix \code{P} for variables with unit variance and scale \code{cov} by the standard deviation of the corresponding predictor
taken from the diagonal of \code{P}}

\item{verbose}{\code{TRUE} or \code{FALSE} to whether printing each lars step}
}
\value{
List with the following elements:
\itemize{
  \item \code{beta}: vector of regression coefficients.
  \item \code{lambda}: penalty of LASSO-type problem for all the sequence of coefficients.
  \item \code{df}: degrees of freedom, number of non-zero predictors at each solution.
  \item \code{sdx}: vector of standard deviation of predictors.
}
The returned object is of the class 'SSI' for which methods \code{fitted} exist. Function \code{plotPath} can be also used
}
\description{
Computes the entire LASSO solution for the regression coefficients, starting from zero, to the
least-squares estimates, via the Least Angle Regression (LARS) algorithm (Efron, 2004). It uses as inputs
a variance matrix among predictors and a covariance vector between response and predictors.
}
\details{
Finds solutions for the regression coefficients in a linear model
\ifelse{html}{\out{<center>y<sub>i</sub> = <b>x</b>'<sub>i</sub> <b>&beta;</b> + e<sub>i</sub></center>}}{\deqn{y_i=\textbf{x}_i'\boldsymbol{\beta}+e_i}{y_i = x'_i beta + e_i}}

where
\ifelse{html}{\out{y<sub>i</sub>}}{\eqn{y_i}{y_i}} is the response for the \ifelse{html}{\out{i<sup>th</sup>}}{\eqn{i^{th}}{i^th}} observation,
\ifelse{html}{\out{<b>x</b><sub>i</sub>=(x<sub>i1</sub>,...,x<sub>ip</sub>)'}}{\eqn{\textbf{x}_i=(x_{i1},...,x_{ip})'}{x_i=(x_i1,...,x_ip)'}}
is a vector of \eqn{p}{p} predictors assumed to have unit variance,
\ifelse{html}{\out{<b>&beta;</b>=(&beta;<sub>1</sub>,...,&beta;<sub>p</sub>)'}}{\eqn{\boldsymbol{\beta}=(\beta_1,...,\beta_p)'}{beta=(beta_1,...,beta_p)'}}
is a vector of regression coefficients, and
\ifelse{html}{\out{e<sub>i</sub>}}{\eqn{e_i}{e[i]}}
is a residual.

The regression coefficients
\ifelse{html}{\out{<b>&beta;</b>}}{\eqn{\boldsymbol{\beta}}{beta}}
are estimated as function of the variance matrix among
predictors (\ifelse{html}{\out{<b>P</b>}}{\eqn{\textbf{P}}{P}}) and the covariance vector between response and predictors (\ifelse{html}{\out{<b>cov</b>}}{\eqn{\textbf{cov}}{cov}}) by minimizing the penalized mean squared error function

\ifelse{html}{\out{<center>-<b>cov</b>' <b>&beta;</b> + 1/2 <b>&beta;</b>'<b>P</b><b>&beta;</b> + 1/2 &lambda; ||<b>&beta;</b>||<sub>1</sub></center>}}{\deqn{-\textbf{cov}' \boldsymbol{\beta} + 1/2 \boldsymbol{\beta}'\textbf{P} \boldsymbol{\beta} + 1/2\lambda||\boldsymbol{\beta}||_1}{-cov' beta + 1/2 beta' P beta + 1/2 lambda ||beta||_1}}

where \ifelse{html}{\out{&lambda;}}{\eqn{\lambda}{lambda}}
is the penalization parameter and
\ifelse{html}{\out{||<b>&beta;</b>||<sub>1</sub> = &sum;|&beta;<sub>j</sub>|}}{\eqn{||\boldsymbol{\beta}||_1=\sum{|\beta_j|}}{||beta||_1=sum(|beta_j|)}}
is the L1-norm.

The algorithm to find solutions for each \ifelse{html}{\out{&beta;<sub>j</sub>}}{\eqn{\beta_j}{beta_j}} is fully described in Efron (2004) in which the "current correlation" between the predictor
\ifelse{html}{\out{x<sub>ij</sub>}}{\eqn{x_{ij}}{xij}}
and the residual
\ifelse{html}{\out{e<sub>i</sub> = y<sub>i</sub> - <b>x</b>'<sub>i</sub> <b>&beta;</b>}}{\eqn{e_i=y_i-\textbf{x}_i'\boldsymbol{\beta}}{e_i = y_i - x'_i beta}}
is expressed (up-to a constant) as

\ifelse{html}{\out{<center>r<sub>j</sub> = cov<sub>j</sub> - <b>P</b>'<sub>j</sub> <b>&beta;</b></center>}}{\deqn{r_j=cov_j-\textbf{P}_j'\boldsymbol{\beta}}{r_j = cov_j - P'_j beta}}

where
\ifelse{html}{\out{cov<sub>j</sub>}}{\eqn{cov_j}{cov_j}}
is the \ifelse{html}{\out{j<sup>th</sup>}}{\eqn{j^{th}}{j^th}} element of
\ifelse{html}{\out{<b>cov</b>}}{\eqn{\textbf{cov}}{cov}} and
\ifelse{html}{\out{<b>P</b><sub>j</sub>}}{\eqn{\textbf{P}_j}{P_j}}
is the \ifelse{html}{\out{j<sup>th</sup>}}{\eqn{j^{th}}{j^th}} column of the matrix
\ifelse{html}{\out{<b>P</b>}}{\eqn{\textbf{P}}{P}}
}
\examples{
  require(SFSI)
  data(wheat.E1) 
  y = as.vector(Y[,"YLD"])  # Response variable
  X = scale(WL[[9]])        # Predictors

  # Training and testing sets
  tst = sample(seq_along(y),ceiling(0.3*length(y)))
  trn = seq_along(y)[-tst]

  # Calculate covariances in training set
  XtX = var(X[trn,])
  Xty = cov(y[trn],X[trn,])
  
  # Run the penalized regression
  fm = lars2(XtX,Xty)  
  fm = lars2(XtX,Xty,method="LAR-LASSO")  
}
\references{
Efron B, Hastie T, Johnstone I, Tibshirani R (2004). Least angle regression. \emph{The Annals of Statistics}, \bold{32}(2), 407–499.

Friedman J, Hastie T, Tibshirani R(2010). Regularization paths for generalized linear models via coordinate descent. \emph{Journal of Statistical Software}, \bold{33}(1), 1–22.

Hastie T, Efron B (2013). lars: least angle regression, Lasso and forward stagewise. \url{https://cran.r-project.org/package=lars}.

Tibshirani R (1996). Regression shrinkage and selection via the LASSO. \emph{Journal of the Royal Statistical Society B}, \bold{58}(1), 267–288.
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos. Adapted from 'lars' package (Hastie & Efron, 2013)
}
\keyword{lars2}
