% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{Methods_SSI}
\alias{fitted.SSI}
\title{SSI methods}
\usage{
\method{fitted}{SSI}(object, ...)
}
\arguments{
\item{object}{An object of the class 'SSI' returned either by 'lars2' or 'solveEN' functions}

\item{...}{Other arguments: a matrix \code{X} of predictors with \code{n} rows and number of columns equal to \code{ncol(object$beta)}}
}
\description{
Returns the predicted values for each value of lambda (in columns) for each row of the provided matrix \code{X} 
}
\examples{
  require(SFSI)
  data(wheat.E1) 
  y = as.vector(Y[,"YLD"])  # Response variable
  X = WL[[9]]               # Predictors

  # Run the penalized regression
  fm = solveEN(var(X),cov(y,X))   
  
  yHat = fitted(fm,X=X)        # Predicted values
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
