% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{scale_crossprod}
\alias{scale_crossprod}
\title{scale_crossprod function}
\usage{
scale_crossprod(XtX)
}
\arguments{
\item{XtX}{Cross-product (\eqn{X'X}) of a matrix \eqn{X} containing \eqn{n} observations and \eqn{p} (centered) variables.}
}
\value{
A list object containing the elements:
\itemize{
  \item 'XtX': squared matrix containing the scaled cross-product matrix \eqn{X'X}
  \item 'sdX': vector containing the standard deviation.
}
}
\description{
Scales the cross-product matrix \eqn{X'X} by dividing each element by its corresponding standard deviations.
The entry corresponding to the cross-product \eqn{x'y} of the (centered) variables 'x' and 'y' is divided by the product of the standard deviations
\eqn{sqrt(x'x)} and \eqn{sqrt(y'y)}. The scaled matrix will have all diagonal entries equal to 1.
}
\examples{
set.seed(1234)
require(SFSI)
# Simulate matrix
n = 100; p=10
X = matrix(rnorm(n*p),ncol=p)
COV = var(X)   # Cross-product X'X

COV2 = scale_crossprod(COV)
COV2$XtX   # equal to: var(scale(X))
COV2$sdX   # equal to: apply(X,2,sd)
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
