% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_utils_wC.R
\name{scale_crossprod}
\alias{scale_crossprod}
\title{scale_crossprod function}
\usage{
scale_crossprod(XtX, void = FALSE)
}
\arguments{
\item{XtX}{Variance-covariance matrix among \eqn{p} variables}

\item{void}{\code{TRUE} or \code{FALSE} to whether return or not return the output.
When \code{FALSE} no result is displayed but the input is modified. Default \code{void=FALSE}}
}
\value{
A squared matrix with the recalculated variances and covariances
}
\description{
Recalculate a variance-covariance matrix among \eqn{p} variables to its equivalent for the variables scaled to have unit variance.
The recalculated matrix will contain as off-diagonal entries the covariance between the scaled variables
\eqn{x*=x/\sigma_x} and \eqn{y*=y/\sigma_y}, formed by dividing original variables \eqn{x} and \eqn{y} by their
standard deviation \eqn{\sigma_x} and \eqn{\sigma_y}, given by
\deqn{cov(x*,y*)=cov(x,y)/(\sigma_x\sigma_y)}
while in the diagonal the variance will be \eqn{var(x*)=var(x)/\sigma^2_x=1}.
}
\examples{
set.seed(1234)
require(SFSI)
# Simulate matrix
n = 100; p=10
X = matrix(rnorm(n*p),ncol=p)
COV = cov(X)   # Cross-product X'X

scale_crossprod(COV)
# it must be equal to:
cov(scale(X))    # or
SD=apply(X,2,sd); cov(scale(X,center=FALSE,scale=SD))

# Using void=TRUE
scale_crossprod(COV,void=TRUE)
COV   # notice that COV was modified
}
\author{
Marco Lopez-Cruz (\email{lopezcru@msu.edu}) and Gustavo de los Campos
}
